Adding in Toby's blur shader demo code.
Split out the shader management code, but not the shaders themselves. Do we
want a shader/ package with the shader.py code in __init__ and actual
shaders in other package modules?

Also pulled out his basic geometric models into a separate module 'cos that
might be handy.

Didn't split out the surface / etc. stuff as I still don't quite get it and
how it relates to the code already in the pyglet.image package.



git-svn-id: http://pyglet.googlecode.com/svn/trunk@73 14d46d22-621c-0410-bb3d-6f67920f7d95