Sorted out the __del__ issue: each time you access a pointer's .contents attribute it creates a new instance for the referenced data. And thus it would free that structure (an FT_FaceRec) when that access went out of scope. The fix was to keep the FT_FaceRec structure around instead of the pointer to it.
It still doesn't work. If you run text_test.py you'll the problem it in
action. Dedent line 29 in pyglet/text/__init__.py so the append is done once
after the loop runs and you'll see what I mean.

To make the error more pronounced, add an "a" to the start of the "WAwa"
test text - you should get a bizzarro "invalid outline" error.

Something in the setting up of the OGL texture is screwing us up. That
worries me since it makes me think of memory corruption badness. I can't
see anything obvious though...

Having said all that, there's most likely better methods of copying the
rendered glyph data to a texture. I'm not sure we'll ever want the
data in anything other than a texture, but to play it safe I'm going via an
Image for now.



git-svn-id: http://pyglet.googlecode.com/svn/trunk@82 14d46d22-621c-0410-bb3d-6f67920f7d95