Some refactoring of the text stuff in preparation of non-freetype rendering.
We no longer return an Image from render_char, but rather a Glyph which has
an OGL texture.

The base module was created to avoid all sorts of nasty circular import
nonsense.

I figured out that I was passing the wrong arg to FT_Get_Kerning (I was
passing the charcode instead of the glyph's char index) but that still
hasn't fixed the kerning. In fact, it's slightly worse now since for
the "Wo" kerning it returns -256 (and 0 for all others)



git-svn-id: http://pyglet.googlecode.com/svn/trunk@89 14d46d22-621c-0410-bb3d-6f67920f7d95