Image regression testing.
Moved all tests from tests/pyglet/* to tests/*.  This was causing many
headaches with relative/absolute imports because of the confusion between
the two pyglet packages.  Also less typing now::

  python tests/test.py image

will run all regression tests on the image section.

There is a bug in the image tests in that they don't all set the codecs
they need to work.  This is annoying to fix because of the stupid codec
API.  Will fix this up later.  In meantime, if an image test fails for
a reason like "DDS header invalid", just re-run that test separately.

If it turns out that images are not loaded/displayed pixel perfect across
platforms we can remove the regression images from the test suite, but
I suspect these ones should be ok.



git-svn-id: http://pyglet.googlecode.com/svn/trunk@233 14d46d22-621c-0410-bb3d-6f67920f7d95