Merging richard-glx-version back into trunk.
I will wait on Anthony's output of tools/info.py before putting in a
special-case which prevents the call to glXDestroyContext for users of ATI
cards with GLX v1.2 (sigh).

There's additional problems with his card running the tests, but I guess
we'll just have to try to deal with them as they arise.


git-svn-id: http://pyglet.googlecode.com/svn/trunk@289 14d46d22-621c-0410-bb3d-6f67920f7d95