Add selection-bar (better name, anyone?)
Fix up slider arrows so they may be recolored, and allowed coloring of
image-based widgets.

Removed the now-unnecessary optimisation of rendering stuff to texture.
image-based buttons with text are still rendered to texture, but that will
change when I implement the new text button code.

As a result of the refactor we can now have rotated text buttons ;)



git-svn-id: http://pyglet.googlecode.com/svn/trunk@1208 14d46d22-621c-0410-bb3d-6f67920f7d95